+++ CHANGELOG.md
@@ -20,9 +20,9 @@
 
 ### Features
 
-* ✨ 集成 tui-calendar ([09f16b6](https://github.com/haydenull/logseq-plugin-calendar/commit/09f16b6f6e8d68527f4bbe5f91a531b1bcf453f4))
-* add setting modal ([5152b47](https://github.com/haydenull/logseq-plugin-calendar/commit/5152b4759bde0ba25e110eb888899ef20f9f4996))
-* show daily log in calendar ([9b6971e](https://github.com/haydenull/logseq-plugin-calendar/commit/9b6971ee0f2d67c88893ba23467efc7fcb083794))
+* ✨ 集成 tui-calendar ([09f16b6](https://github.com/haydenull/logseq-plugin-agenda/commit/09f16b6f6e8d68527f4bbe5f91a531b1bcf453f4))
+* add setting modal ([5152b47](https://github.com/haydenull/logseq-plugin-agenda/commit/5152b4759bde0ba25e110eb888899ef20f9f4996))
+* show daily log in calendar ([9b6971e](https://github.com/haydenull/logseq-plugin-agenda/commit/9b6971ee0f2d67c88893ba23467efc7fcb083794))
 
 # 1.0.0 (2022-01-18)